Make Java work again by reducing its maximum heap size.
Otherwise, it will crash with an out-of-memory error on startup
because it tries to allocate 1.05 G. While one option is to increase
our memory limits, a better one is to make it grab less RAM. The
startup code's idea of "a reasonable fraction" of total physical
memory in the machine is not our idea of a reasonable amount for one
process on scripts to use.

Reported by tabbott, debugged by quentin and geofft, fix suggested by
mitchb.



git-svn-id: file:///mit/scripts/svn/trunk@1168 db9d59ff-b01e-0410-9b1a-cd9a8c06840f