execsys: Have binfmt_misc detect executable PHP scripts via magic, not extension.
For PHP scripts run over the web, suexec runs php explicitly (and
regardless of the .php file's executable bit), so this change isn't
relevant. A recent kernel change reordered the binfmt_script and
binfmt_misc priority; this allows executable PHP scripts that begin with
<?php to continue being run via binfmt_misc, but those that begin with
#! to use the interpreter they specify.

PHP scripts that don't begin with <?php are assumed to be web pages, not
scripts you'd explicitly exec. This is a slightly smaller hammer than
removing the binfmt_misc line for PHP entirely, but really, you should
be using a #! line if you intend to be able to exec your PHP script.



git-svn-id: file:///mit/scripts/svn/trunk@1338 db9d59ff-b01e-0410-9b1a-cd9a8c06840f