Speed up postfix acceptance of mail to foo@scripts.mit.edu
Postfix does virtual alias expansion on *all* addresses, both
virtual/remote and local, which means that even mail to a
normal scripts user will be looked up in not only the regular
aliases file, but the virtual maps we set up.  If we don't
match those addresses in either the hashed virtual file or
the regexp virtual file, then an LDAP query will occur instead
of quickly accepting the RCPT address.  Short-circuit that
by matching *@scripts.mit.edu as a regexp and mapping it with
no change.  If we rewrite it as the unqualified user, then
a recursive virtual match happens (all virtual rewrites are
subject to recursive matching).  Simple, eh?


git-svn-id: file:///mit/scripts/svn/trunk@1452 db9d59ff-b01e-0410-9b1a-cd9a8c06840f