Backport cluster-glue, heartbeat, and pacemaker from F12
We want to use pacemaker for hacron.  It's presently available in
F12, but not F11.  In F12, it's built against heartbeat 3, which
is a completely different beast than the heartbeat 2 available in
F11, so we need to backport that as well.  heartbeat 3, in turn,
depends on cluster-glue, which is not presently in F11.  All
three of these build cleanly with no source changes so long as
their dependencies are present.


git-svn-id: file:///mit/scripts/svn/trunk@1459 db9d59ff-b01e-0410-9b1a-cd9a8c06840f