__scripts/needcerts: Add support for working around Safari
Safari on Mac OS X (or more properly, CFNetwork and the rest of the SSL
stack) doesn't properly support SSLVerifyClient Optional, which is our
default for :444. In particular, if you don't have an identity
preference set, only SSLVerifyClient Require will trigger the dialog to
set an identity preference and present a certificate to the site:

http://lists.apple.com/archives/apple-cdsa/2009/Apr/msg00041.html

We can work around this by checking for the Safari user-agent in
/__scripts/needcerts and renegotiating SSLVerifyclient Require. Forcing
the Require behavior on Safari users that reach this page is reasonable
because this page is only (supported to be) reached as an ErrorDocument
401; if you're intentionally using AuthOptional on to take advantage of
the optional authentication, you'll never trigger the 401 error.



git-svn-id: file:///mit/scripts/svn/trunk@1483 db9d59ff-b01e-0410-9b1a-cd9a8c06840f