Temporarily scriptsify shadow-utils to fix max length of group names
The upstream shadow-utils package used to hardcode the max length of
a group name to 16 characters, and Fedora patched it to make that
length match utmpx.ut_name's length (32) and documented it as being 32.
The last upstream update that Fedora took changed the code to default
to 16, but allow a nondefault value via a --with flag to configure.
Fedora's patch thus no longer applied and they dropped the relevant
hunk, but neglected to add the --with flag in the spec file, so the
max group name length dropped back to 16, causing packages that install
groups with names longer than 16 characters to begin failing to install
properly mid-release.  We'll be pursuing this issue with Fedora.


git-svn-id: file:///mit/scripts/svn/trunk@1507 db9d59ff-b01e-0410-9b1a-cd9a8c06840f