Make procmail and cron try harder to actually mail the signup user
At present, we try to set up the initial crontab or procmailrc to
send mail to the user who signed a locker up for the cron or mail
service by stripping off the @REALM from the Kerberos principal.
We need to also strip non-null instances off to get the username.
(If you're doing signups for lockers using machine or service
principals, you ought to be able to fend for yourself.)


git-svn-id: file:///mit/scripts/svn/branches/locker-dev@1534 db9d59ff-b01e-0410-9b1a-cd9a8c06840f