Revert r1510.  Back out to OpenAFS 1.4.11.
1.4.12 may not have been as much of a Nice Thing (TM) as we'd hoped.
Since upgrading to it, our servers have been flapping on a more-than-daily
basis, almost always due to bogus AFS lost contact messages, resulting
in loads shooting up into the hundreds while the machines are stuck
in IOwait.  When this includes the server that contrib.scripts is on,
the machine falls out of the pool.  In all cases, it results in a
service slowdown.  Sometimes a reboot has been necessary because 'fs checks'
doesn't always clear it up.  While we haven't found a conclusive cause
for the problems, they seem to point to AFS, and began happening the
same night we upgraded to 1.4.12.


git-svn-id: file:///mit/scripts/svn/trunk@1545 db9d59ff-b01e-0410-9b1a-cd9a8c06840f