Fix whoisd bug introduced in r1741
The code in r1741 would cause the message about having trouble
looking up LDAP records to be returned for nonexistent host
records as well as for actual LDAP problems.  If we get an empty
search result back, but don't experience an error, we're done
and don't need to keep trying.


git-svn-id: file:///mit/scripts/svn/trunk@1742 db9d59ff-b01e-0410-9b1a-cd9a8c06840f