Move 389-ds's slapd-scripts.socket to /var/run
It turns out that mode 777 directories containing files
that daemons use is... not the most brilliant thing we've
done.  389-ds has finally decided to insist on clobbering
the permissions of /var/run/dirsrv to be less foolish,
but several of our daemons and client programs need to
be able to access the LDAP daemon's socket.  Come visit
it in its new home, conveniently located just two directories
below the root.


git-svn-id: file:///mit/scripts/svn/trunk@1818 db9d59ff-b01e-0410-9b1a-cd9a8c06840f