ldapize: Give a custom error for unknown git vhosts
The previous commit changed the error message on unknown git vhosts
from this (caused by a mangled Perl backtrace):
  fatal: protocol error: bad line length character: Can'
to the message provided by upstream:
  fatal: The remote end hung up unexpectedly

This further changes the error message to the custom
  fatal: remote error: No such host nonexistent.scripts.mit.edu

I do not personally support this commit, because this just isn’t a
common problem, and using the upstream message was a good enough
improvement.  However, it is included by demand.

git-svn-id: file:///mit/scripts/svn/trunk@1824 db9d59ff-b01e-0410-9b1a-cd9a8c06840f