Update OpenAFS from 1.4.14 to 1.6.0pre6
kaduk finished porting the openafs-scripts patch to 1.6.0, which was
then lightly modified be me to get it to compile -- mostly fixing
typos.

The openafs.spec.patch was also fixed, including another chmod +x hack
on some libraries to get rpm to properly process them for Provides
directives.  Interestingly, these libraries had the correct mode in
1.4.14.

The patches that were added to get 1.4.14 to compile have also been
removed, as 1.6.0pre6 compiles on kernel 2.6.38 unmodified.

openafs-numsysnames is getting removed because 32 is now the default.

openafs-localcsdb is getting removed because the change was included
in kaduk's patchset.



git-svn-id: file:///mit/scripts/svn/branches/fc15-dev@1913 db9d59ff-b01e-0410-9b1a-cd9a8c06840f