Unbork yum-plugin-post-transaction-actions
We tried to be too clever for ourselves; stop doing that.  Our
intent was to make sure that permission changes were made on both
installs and updates of packages containing relevant files, but
we didn't want to duplicate each line, so we chose to trigger
on "any" instead of "install" and "update".  An update actually
consists of installation of one package and removal of its previous
version, and filelists aren't available to yum or rpm for removed
packages, so if your trigger is a filename on a removed package,
the plugin fails ungracefully.  Since an update includes an install
of the new package version, the "install" package set includes all
updated packages, and we should only trigger on it.  Additionally,
the -f option we gave chmod to blow through errors silently when
files weren't there (for removal of a package) won't be necessary
as a result.


git-svn-id: file:///mit/scripts/svn/trunk@1926 db9d59ff-b01e-0410-9b1a-cd9a8c06840f