Fix syslog-ng/systemd interaction, or "socket to me!"
systemd creates and passes datagram sockets to daemons (at least by
default).  Fedora's default configuration for syslog-ng expected the
socket passed in to be a stream socket (and we slurped that in forever
ago).  When syslog-ng sees that the socket it received isn't of the
type it expected, instead of either coping or barfing, it just opens
a new /dev/log socket of the type it wants, which is just hunky-dory
for programs that connect after it disowns systemd's socket, but for
anything that already connected to the datagram socket from systemd,
they're logging to a useless socket that systemd isn't poll()ing, and
thus can block (and, depending on what programs end up in this situation,
potentially wedge your entire system).  So, uh... let's not do that
anymore.  Fedora's also changing their default config.

See RHBZ #742624


git-svn-id: file:///mit/scripts/svn/trunk@2100 db9d59ff-b01e-0410-9b1a-cd9a8c06840f