Correct picker.mit.edu aliases
In r2106, there was a change to the aliases line that added pickr and
pickr.mit.edu, bringing this configuration into sync with LDAP.
However, this is clearly Wrong (TM) as the certificate has been issued
for picker.mit.edu only.  This change reflects a re-sync with LDAP
after the aliases have been corrected.


git-svn-id: file:///mit/scripts/svn/trunk@2107 db9d59ff-b01e-0410-9b1a-cd9a8c06840f