Make whoisd (hopefully) more reliable
whoisd was systemd-ified for F15, replacing the cron-based @reboot
startup. Unfortunately, this made whoisd prone to failing to start, as
the dependencies were not correctly stated. This adds dirsrv as a
dependency, as it was implicitly in the cron-based system.

Additionally, prevent twistd from making a /twistd.pid by specifying
the path that was used from the cron-start era. The --pidfile was
removed as it was not needed by systemd, which uses cgroups to manage
processes. Unfortunately, twistd insists on pidfiles, so this prevents
twistd from polluting /.



git-svn-id: file:///mit/scripts/svn/trunk@2114 db9d59ff-b01e-0410-9b1a-cd9a8c06840f