Create patch for current openafs-stable-1_6_x as fake 1.6.1.1
Create a single large patch containing the diff between the 1.6.1
release and a new openafs-stable_1_6_1_1-branch, which was made from a
rebase of openafs-stable-1_6_x onto openafs-stable-1_6_1-branch. We
continue to revert the things reverted in 1.6.1, including "offline-less
vos release". This patch supersedes three of the patches we previously
cherry-picked.

This allows us to compile on Linux 3.6.

In order to make this work, we run ./regen.sh (= autoreconf) in the
build process. Consider removing this once we get back to closely
tracking an actual OpenAFS release.

The actual git branch is in openafs-stable-1_6_1_1-branch on
https://github.com/sipb/openafs (also in w-e:/home/scripts-build/openafs).



git-svn-id: file:///mit/scripts/svn/trunk@2335 db9d59ff-b01e-0410-9b1a-cd9a8c06840f