Remove InCommon Root CA from certificate chain
MIT recently switched to the InCommon CA, rather than the MIT CA,
which resulted in Scripts serving intermediate certificates.
Unfortunately, the instructions from InCommon had the certificates in
the incorrect order, and also included the Root CA.  The Root CA is
unnecessary for a server to be sending, so removing it reduces the SSL
handshake size, as well as correcting the certificate order.


git-svn-id: file:///mit/scripts/svn/trunk@2351 db9d59ff-b01e-0410-9b1a-cd9a8c06840f